fix: mudando a arquitetura do projeto
