+++ api/services/ProfessorAdmin/GenerateToken.js
@@ -1,0 +1,11 @@
+const jwt = require("jsonwebtoken");
+const secret = process.env.prof_secret;
+
+const generate_token = (id) => {
+    return jwt.sign({ id }, secret, {
+        expiresIn: '7d',
+    });
+};
+
+
+module.exports = generate_token;
